Package MovableType export file parser
Related to #10

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/yamadatt/mt2hugo/issues/10?shareId=XXXX-XXXX-XXXX-XXXX).
